fix:  link install test #129 (#140)

+++ src/inject.ts
@@ -7,6 +7,7 @@ import {
   getInputMap,
   getInputMapUrl,
   getResolutions,
+  inputMapExists,
   startLoading,
   stopLoading,
 } from './utils'
@@ -35,9 +36,13 @@ export default async function inject(
   htmlFile: string,
   packages: string[],
   flags: Flags,
-) {
+): Promise<string> {
   const inputMap = await getInputMap(flags)
   const env = getEnv(flags, true, inputMap)
+  if (!(await inputMapExists(flags))) {
+    console.error('No input map found, nothing to inject.')
+    return ''
+  }
   if (!(await htmlExists(htmlFile))) {
     console.warn(`Warning: HTML file ${htmlFile} does not exist, creating one`)
     await fs.writeFile(htmlFile, defaultHtmlTemplate, 'utf-8')
